requirements.txt
# --- Core Web Framework ---
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-WTF>=1.2.1
Flask-Limiter>=3.6.0
Flask-Migrate>=4.0.5
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.29

# --- Env Vars ---
python-dotenv>=1.0.1

# --- AI SDKs ---
openai~=0.27.10
anthropic>=0.36.0    # 'generator.py'에서 사용되는 Claude 라이브러리 가정

# --- PostgreSQL Driver ---
psycopg2-binary>=2.9.9

# --- Core Dependencies (Flask/SQLAlchemy Internal) ---
werkzeug~=3.1.3
alembic~=1.13.1      # Flask-Migrate의 내부 마이그레이션 엔진

# --- (Optional but Recommended for Production) ---
gunicorn>=22.0.0     # 리눅스 배포 시 WSGI 서버
redis>=5.0.0         # Flask-Limiter 저장소용 (운영 시)
dotenv~=0.9.9
bleach~=6.3.0
jsonschema~=4.25.1
itsdangerous~=2.2.0
google~=3.0.0

requests>=2.32.0
python-dateutil>=2.9.0
